pyproject.toml
[tool.poetry]
name = "onekid"
version = "0.1.0"
description = ""
authors = ["mortilele <ali.mars.99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-model-utils = "^4.1.1"
djangorestframework = "^3.12.4"
django-3-jet = "^1.0.8"
drf-spectacular = "^0.17.2"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.1"
gevent = "^21.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
